// apps/web/src/api/generated/model/todayAttendanceResponseDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */
import type { PunchDetailDto } from './punchDetailDto';
import type { TodayAttendanceResponseDtoAttendanceState } from './todayAttendanceResponseDtoAttendanceState';
import type { TodayGeofenceDto } from './todayGeofenceDto';
import type { TodayShiftDto } from './todayShiftDto';

export interface TodayAttendanceResponseDto {
  /** @nullable */
  shift: TodayShiftDto | null;
  attendanceState: TodayAttendanceResponseDtoAttendanceState;
  /** @nullable */
  checkIn: PunchDetailDto | null;
  /** @nullable */
  checkOut: PunchDetailDto | null;
  workingDurationSeconds: number;
  /** @nullable */
  geofence: TodayGeofenceDto | null;
  canCheckIn: boolean;
  canCheckOut: boolean;
  warnings: string[];
}